add handlers with repos

+++ backend/models/game.go
@@ -1,7 +1,7 @@
 package models
 
 type Game struct {
-	ID          int64   `json:"id" db:"id"`
+	ID          int     `json:"id" db:"id"`
 	Title       string  `json:"title" db:"title" binding:"required"`
 	Description string  `json:"description" db:"description" binding:"required"`
 	Price       float64 `json:"price" db:"price" binding:"required"`
